CHANGE: Started Using Lazy.nvim
Started using the Lazy.nvim plugin as an alternative to packer. Although
Packer was great, the future seems to be the Lazy.nvim plugin which
includes several great features (such as ez lazy loading and GUI for
installation of certain plugins).

To install this, one must first remove packer. This can easily be done
by removing the data from the ~/.local/nvim folder together with the
saved data inside of the plugin directory of ~/.config/nvim directory.